pyproject.toml
[tool.poetry]
name = "test-razorpay"
version = "0.1.0"
description = ""
authors = ["anand2312 <anandkrishna2312@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
Jinja2 = "^3.0.1"
uvicorn = "^0.15.0"
razorpay = "^1.2.0"
python-multipart = "^0.0.5"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
